Make State generic: authentication and connection methods
